Move MeteorStubs plugin under "internal" subdirectory.
