+++ npm-packages/meteor-vite/src/meteor/IPC/methods/server-development.ts
@@ -5,7 +5,7 @@ import Logger from '../../../utilities/Logger';
 import MeteorEvents, { MeteorIPCMessage } from '../MeteorEvents';
 import { MeteorViteConfig } from '../../../MeteorViteConfig';
 import { MeteorStubs } from '../../../plugin';
-import { ProjectJson } from '../../../plugin/stubs/MeteorStubs';
+import { ProjectJson } from '../../../plugin/internal/MeteorStubs';
 import { RefreshNeeded } from '../../../ViteLoadRequest';
 import CreateIPCInterface, { IPCReply } from '../interface';
