+++ npm-packages/meteor-vite/src/plugin/internal/MeteorStubs.ts
